pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proppilot"
version = "0.1.0"
description = "Airbnb Property Management Automation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "jinja2>=3.1",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "apscheduler>=3.10,<4",
    "icalendar>=5.0",
    "imap-tools>=1.5",
    "httpx>=0.25",
    "twilio>=8.10",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.6",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "httpx",  # for FastAPI TestClient
]

[project.scripts]
proppilot = "proppilot.app:main"

[tool.setuptools.packages.find]
where = ["src"]
